pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-processor"
authors = [
    {name = "PDF处理器开发团队", email = "support@pdfprocessor.com"},
]
description = "一个功能强大的PDF处理工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
dynamic = ["version"]
dependencies = [
    # 核心PDF处理依赖
    "PyPDF2>=1.26.0",
    "reportlab>=3.5.0",
    # GUI界面依赖
    "PyQt5>=5.15.0",
    # 图像处理依赖
    "Pillow>=8.0.0",
    "pdf2image>=1.14.0",
    # OCR相关依赖
    "pytesseract>=0.3.7",
    # 数字签名相关依赖
    "PyHanko>=0.12.0",
    "pyhanko-certvalidator>=0.12.0",
    "cryptography>=3.4.8",
    # 文档格式转换依赖
    "ebooklib>=0.17.1",
    "python-docx>=0.8.11",
    "openpyxl>=3.0.7",
    "pandas>=1.3.0",
    "python-pptx>=0.6.18",
    # 性能和系统相关依赖
    "psutil>=5.8.0",
    "numpy>=1.20.0",
    "concurrent-log-handler>=0.9.20",
    # 视觉差异检测相关依赖
    "opencv-python>=4.5.0",
]

[project.optional-dependencies]
# 开发环境依赖
dev = [
    "pytest>=6.0.0",
    "pytest-qt>=4.0.0",
    "flake8>=3.8.0",
    "black>=21.0.0",
    "mypy>=0.812",
    "coverage>=5.5.0",
    "sphinx>=4.0.0",
    "pylint>=2.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "ui*", "utils*", "config*", "plugins*"]
